pyproject.toml
[project]
name = "asynccdp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cdpkits",
    "pydantic>=2.11.4",
]

[tool.ruff.lint]
preview = true
select = ["E", "F", "W", "I", "UP"]
exclude = ["tests", "tests/*", "demos/*"]

[tool.ruff.format]
preview = true
quote-style = "single"

[dependency-groups]
ruff = [
    "ruff>=0.11.6",
]

[tool.ruff]
line-length = 120
target-version = "py313"
